fix: support different formmethod on submitter button
Remix's `<Form/>` or more precisely the [`FormImpl`][1] and it's
[`useSubmit` (`useSubmitImpl`)][2] hook used to ignore the potential
existance of a submitter responsible of the form to be submitted. This
meant that using `formmethod` or `formaction` attributes on a submit
button to override the default method or action of the form wasn't
possible.

This has been improved in:
  - https://github.com/remix-run/remix/commit/3f04b8b5b3d9e9aaf61b0d6923139eacd10dda1e
  - https://github.com/remix-run/remix/commit/c03d0fd0efa4aef57e79c6811eadee5383200e24

There's however still a pending PR on Remix's side, which would fix
Remix's `<Form/>` component itself:
  - https://github.com/remix-run/remix/pull/2162

However, since Remix 1.6.1, this issue can already be fixed for
`<ValidatedForm/>` which uses Remix's `useSubmit` (`useSubmitImpl`) and
already passes the `nativeEvent.submitter` – if any –  and thus allow
the attributes `formmethod` & `formaction` to be used if they are set.

For this, and similarly to the PR mentioned above, `<ValidatedForm/>`
must not override the `method` in the options passed to the `submit()`
function. Instead, inferring which method and action must be used is
left to Remix's `useSubmitImpl`.

FYI: `formenctype` should be supported too. Just be aware that at the
time of this commit, this attribute is not documented as being fully
supported by some browsers (e.g Safari).

[1]: https://github.com/remix-run/remix/commit/3f04b8b5b3d9e9aaf61b0d6923139eacd10dda1e#diff-594d4b6ed5bd4abff7e9319566b5505b6a85753e665ff4d2113e888384a98377R929-R930
[2]: https://github.com/remix-run/remix/commit/c03d0fd0efa4aef57e79c6811eadee5383200e24#diff-594d4b6ed5bd4abff7e9319566b5505b6a85753e665ff4d2113e888384a98377L1095-R1114