Added `infer` type helper
When doing basic form validation, pretty much everything can be done using just `zfd`, however if you want a portable version of the type from that schema you have to reach for zod.

That is a pain so I'm adding this export of the zod infer type as a convenience for developers.